BUG#21776505: GR DOES NOT HANDLE BIG GTID_EXECUTED CORRECTLY
There were several issues on how big GTID_EXECUTED sets were handled
on Group Replication pipeline, in particular when GTID_EXECUTED size
exceeded the 65535 characters length. The issues were:
  1) snapshot_version length computation was incorrectly limited to
     65535 characters length on Transaction_context_log_event;
  2) snapshot_version contained on certification info, which is
     serialized on View_change_log_event, was incorrectly limited to
     65535 characters length;
  3) internal interface between Group Replication plugin and server
     was incorrectly limiting GTID_EXECUTED to 65535 characters
     length.
  4) Column Transactions_committed_all_members column on
     performance_schema.replication_group_member_stats was limited
     to 65535 characters length.

This patch fixes the 4 above issues. Since issue 2) requires the
change of the serialization of View_change_log_event, this patch
makes Group Replication version 0.6.0 incompatible with the previous
ones. Incompatibility rules were added and also a test (with
simulated versions) to validate them.

+++ include/mysql/group_replication_priv.h
@@ -125,7 +125,7 @@ void get_server_startup_prerequirements(Trans_context_info& requirements,
   @param[out] length                binary string length
 */
 bool get_server_encoded_gtid_executed(uchar **encoded_gtid_executed,
-                                      uint *length);
+                                      size_t *length);
 
 #if !defined(DBUG_OFF)
 /**
@@ -138,7 +138,7 @@ bool get_server_encoded_gtid_executed(uchar **encoded_gtid_executed,
 
   @return a pointer to text representation of the encoded set
 */
-char* encoded_gtid_set_to_string(uchar *encoded_gtid_set, uint length);
+char* encoded_gtid_set_to_string(uchar *encoded_gtid_set, size_t length);
 #endif
